Remove is_moderator from 0047 and moved it to new migration 0048.
