+++ database/entity/ContributionMessage.ts
@@ -1,1 +1,1 @@
-export { ContributionMessage } from './0047-messages_tables/ContributionMessage'
+export { ContributionMessage } from './0048-add_is_moderator_to_contribution_messages/ContributionMessage'
